Correct typing errors in properties.hpp and usermodel.hpp
Corrected type of int for vent_rate_flag (was doubple in line 2176) in
UserModel.hpp and converted map.size() output from size_t to int through
static type casting in Properties.hpp line 192

+++ IsoModel/src/Properties.hpp
@@ -189,7 +189,7 @@ public:
    */
   int size() const
   {
-    return map.size();
+    return static_cast<int>(map.size());
   }
 
 };
